Update plugin.
* Rename Shotgun/ShotGrid to Flow Production Tracking
* Remove obsolete plugin code
